+++ packages/realm/src/index.ts
@@ -105,6 +105,8 @@ export {
   LogLevel,
   mapTo,
   MapToDecorator,
+  MetadataMode,
+  Metadata,
   MigrationCallback,
   MongoDB,
   MongoDBCollection,